requirements.txt
tensorflow==2.12.0
flask==2.3.2
pillow==9.5.0
numpy==1.23.5
matplotlib==3.7.1
tensorflow-datasets
